+++ src/schrage/jobs.rs
@@ -1,4 +1,6 @@
+// pub mod jobs;
 use std::{cmp::Ordering, fmt, vec};
+
 
 #[derive(Copy, Clone, Debug)]
 pub struct Job {
